LayoutTests/media/video-can-play-type.html

<html>
    <head>
        <style>
            video { background-color: yellow; width: 320px; height: 240px;} 
        </style>
    </head>
    <body>

        <video controls ></video>

        <p>Test HTMLMediaElement <em>canPlayType()</em> method.</p>

        <script src=video-test.js></script>
        <script>
        
            video = mediaElement = document.getElementsByTagName('video')[0];

            testExpected("video.canPlayType()", "");
            testExpected("video.canPlayType('video/')", "");
            testExpected("video.canPlayType('video/blahblah')", "");
            testExpected("video.canPlayType('video/mp4; Codecs=\"avc1.4D400C\"')", "probably");
            testExpected("video.canPlayType('        Video/MP4 ; CODECS=\"mp4v.20.8, mp4a.40.2\"')", "probably");
            testExpected("video.canPlayType('audio/mpeg')", "maybe");
            testExpected("video.canPlayType('audio/Wav')", "maybe");
            testExpected("video.canPlayType('video/blahblah; codecs=\"blah,  , blah\"')", "");
            testExpected("video.canPlayType('video/blahblah; codecs=blah')", "");
            testExpected("video.canPlayType('video/blahblah; codecs=\"blah\"')", "");
            testExpected("video.canPlayType('video/blahblah; codecs=\"badcontent')", "");
            testExpected("video.canPlayType('video/blahblah; codecs=badcontent\"')", "");
            testExpected("video.canPlayType('video/blahblah; codecs=&quot;badcontent&quot;')", "");

            endTest();
        </script>
    </body>
</html>
